requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
python-multipart==0.0.9
spacy==3.7.2
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2
streamlit==1.32.0